Switched to real Resque and fakeredis.
It didn't make sense to copy / rewrite *all* of Resque, just to add some
assertions to it. Instead, I'm using fakeredis and actual Resque. That
way, all the persistence happens in-memory (so you don't need to run
Redis yourself), and I can take advantage of the built-in methods Resque
has for looking through queues and running jobs. This should be way more
reliable and easy to maintain.

Currently, the regular assertions pass. I'm still working on the
scheduler assertions.